requirements.txt
# Web UI
streamlit>=1.30.0

# Deep Learning & VLM Core
transformers>=4.46.0
accelerate>=0.26.0
qwen-vl-utils>=0.0.8
pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.10.0.84
easyocr>=1.7.1

# Hybrid RAG System
sentence-transformers>=2.3.0
chromadb>=0.4.22
rank_bm25>=0.2.2

# NOTE: PyTorch should ideally be installed manually to match the system's CUDA version.
# Example for CUDA 12.x: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu124
